NAV-57 Refactor fix circular import.
